Header and footer added
Header contains basic example options.
When opened in phone, the navigation options disappear and become
a button which opens a drawer. This is subject to changes, but shows
responsiveness of the website
Footer is a basic box with a container grid which has lorem ipsum.
Content shall be added there.

Closes #1.