Updated appbar, added calendar
Updated appbar to the proposed design. Shall be up for changes.
The appbar is swipeable on phones and the menu structure is different
based on screen width, which can be experimented with.
Calendar has been added and has been added into home page to preview.
In home page, it shows the upcoming events only, I have added 3 events
of september 2nd for example, but they disappear as the time interval
runs out as it shows only upcoming. Full calendar can be seen using
that link. Home page has basic preview with basic content

Closes #7, Closes #3